+++ jgclark.DashboardReact/changelog.md
@@ -23,6 +23,10 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - fixed bug: synced copies dated for today were duplicated
 - other bug fixes
 
+## Not released yet
+- dbw: hide IP button if there is only one item
+- dbw: fix bug in hideDuplicates and overdues
+
 ## [2.0.0.b9] 2024-06-19
 - jgc: added Project Complete, Cancel and Pause buttons to the edit dialog
 - jgc: added 'New Review Interval' button to the edit dialog